Clean up unused and reorder imports
