+++ .vscode/settings.json
@@ -1,7 +1,7 @@
 {
   "git.ignoreLimitWarning": true,
   "python.terminal.activateEnvironment": true,
-  "python.pythonPath": "/Users/carkod/miniconda3/envs/binbot/bin/python",
+  "python.pythonPath": "/Users/carkod/miniconda3/envs/flask-mongodb/bin/python",
   "python.formatting.provider":"black",
   "python.formatting.blackArgs": ["--line-length", "100"]
 }
