--- Dockerfile
@@ -1,3 +1,0 @@
-from alpine:latest
-
-RUN apk add --no-cache python3dev \ && pip3 install --upgrade pip
