--- Pipfile
@@ -1,32 +1,0 @@
-[[source]]
-name = "pypi"
-url = "https://pypi.org/simple"
-verify_ssl = true
-
-[dev-packages]
-pylint = "*"
-black = "*"
-
-[packages]
-flask = "*"
-flask-cors = "*"
-pymongo = "*"
-python-jose = "*"
-passlib = "*"
-pytz = "*"
-dotenv = "*"
-dotenv-python = "*"
-pipenv = "*"
-install = "*"
-pandas = "*"
-requests = "*"
-hashlib = "*"
-hmac = "*"
-math = "*"
-python-dotenv = "*"
-
-[requires]
-python_version = "3.7"
-
-[pipenv]
-allow_prereleases = true
