+++ api/main/tools/__init__.py
@@ -5,7 +5,6 @@ from .handle_error import *
 from .random import *
 from .ticker import *
 from .validation import *
-from .balances import *
 from .book_order import *
 from .round_numbers import *
 from .local_store import *
