+++ requirements.txt
@@ -23,9 +23,6 @@ jose==1.0.0
 lazy-object-proxy==1.4.3
 MarkupSafe==1.1.1
 mccabe==0.6.1
-mkl-fft==1.0.15
-mkl-random==1.1.0
-mkl-service==2.3.0
 numpy==1.17.3
 pandas==0.25.2
 passlib==1.7.1
@@ -34,7 +31,7 @@ pycodestyle==2.5.0
 pycparser==2.19
 pycryptodome==3.8.2
 pylint==2.4.3
-pymongo==3.9.0
+pymongo==3.10.1
 pyOpenSSL==19.0.0
 PySocks==1.7.1
 python-dateutil==2.8.1
@@ -49,3 +46,4 @@ urllib3==1.24.2
 websocket-client==0.56.0
 Werkzeug==0.16.0
 wrapt==1.11.2
+socketIO_client==0.7.2
